change n_tofbins argument to short in cuda sino projectors

+++ pyparallelproj/wrapper.py
@@ -159,7 +159,7 @@ if os.path.exists(lib_parallelproj_cuda_fname):
                                                                ar_1d_single,      # sigma tof
                                                                ar_1d_single,      # tofcenter_offset
                                                                ctypes.c_float,    # n_sigmas 
-                                                               ctypes.c_int,      # n_tofbins
+                                                               ctypes.c_short,    # n_tofbins
                                                                ctypes.c_int,      # threads per block
                                                                ctypes.c_int]      # number of devices 
   
@@ -176,7 +176,7 @@ if os.path.exists(lib_parallelproj_cuda_fname):
                                                                 ar_1d_single,      # sigma tof
                                                                 ar_1d_single,      # tofcenter_offset
                                                                 ctypes.c_float,    # n_sigmas 
-                                                                ctypes.c_int,      # n_tofbins
+                                                                ctypes.c_short,    # n_tofbins
                                                                 ctypes.c_int,      # threads per block
                                                                 ctypes.c_int]      # number of devices
